requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
asyncpg==0.29.0
pillow==10.2.0